requirements.txt
flask>=3.0.0
werkzeug>=3.0.0
pymupdf>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pillow>=10.0.0
pytesseract>=0.3.10
gunicorn>=21.2.0
